requirements.txt
# Telegram
aiogram>=3.13,<4.0

# Web framework
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32
jinja2>=3.1
python-multipart>=0.0.12

# Database
sqlalchemy>=2.0,<3.0
asyncpg>=0.30
alembic>=1.14

# Queue & Cache
celery[redis]>=5.4
redis>=5.0

# AI
openai>=1.50

# Google Calendar
google-api-python-client>=2.150
google-auth-oauthlib>=1.2
google-auth-httplib2>=0.2

# Config & Validation
pydantic>=2.10
pydantic-settings>=2.6

# Scheduling & Time
pytz>=2024.0

# HTTP
httpx>=0.27
aiohttp>=3.9

# Logging
loguru>=0.7

# Testing
pytest>=8.0
pytest-asyncio>=0.24
aiosqlite>=0.20
